extract PersonInfo to common models
