+++ src/app/features/common/models.ts
@@ -1,0 +1,9 @@
+export interface PersonInfo {
+  id: string;
+  email?: string;
+  name: string;
+  phone?: string;
+  altPhone?: string;
+  address?: string;
+  code?: string;
+}
